+++ lib/tools/text_tools/markdown_preview/markdown_preview_view.dart
@@ -58,6 +58,7 @@ class _MarkdownPreviewViewState extends State<MarkdownPreviewView> {
                 child: Consumer(builder: (context, ref, _) {
                   return Markdown(
                     data: ref.watch(_inputValue),
+                    selectable: true,
                   );
                 }),
               ),